refactor(baseline): extract scoped-inputs.ts to keep create.ts under the large-file threshold
The self-guardrail (dxkit dogfooding its own gate) caught opt 1's additions
pushing src/baseline/create.ts over the 500-line large-file threshold — a
net-new finding vs origin/main. Per the dogfood rule, fix the finding, don't
refresh the baseline.

Extract the scope-aware producer-context input gathering (test-gaps,
hygiene, raw secrets, inline annotations + the empty-hygiene constant) into
src/baseline/scoped-inputs.ts — a cohesive unit, mirroring the gate-cache.ts
extraction in 2.13.3. create.ts drops to 495 lines; no behavior change.
